Changed how the UI system renders quads

+++ crates/fonts/src/char.rs
@@ -1,6 +1,6 @@
 // A font character
 pub struct FontChar {
     pub id: u8,
-    pub min: veclib::Vector2<u32>,
-    pub max: veclib::Vector2<u32>,
+    pub min: veclib::Vector2<u16>,
+    pub max: veclib::Vector2<u16>,
 }
